Refactor repeating sections into their own file
Improves readability
